overall refactor
